refactor: Rename pure-variant to pure-variants
