+++ frontend/src/app/settings/integrations/pure-variants/pure-variants.component.css
